+++ test/unit/mongo_db_admin/main.py
@@ -44,6 +44,7 @@ class ArgParser(object):
         arg_require
         get_val
         arg_xor_dict
+        arg_parse2
 
     """
 
@@ -75,6 +76,7 @@ class ArgParser(object):
         self.arg_valid_val2 = True
         self.opt_con_req = None
         self.opt_con_req2 = True
+        self.argparse2 = True
 
     def arg_cond_req(self, opt_con_req):
 
@@ -188,6 +190,18 @@ class ArgParser(object):
 
         return self.opt_xor_val2
 
+    def arg_parse2(self):
+
+        """Method:  arg_parse2
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_parse2.
+
+        Arguments:
+
+        """
+
+        return self.argparse2
+
 
 class ProgramLock(object):
 
@@ -224,6 +238,8 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
+        test_arg_parse2_false
+        test_arg_parse2_true
         test_help_true
         test_help_false
         test_arg_req_false
@@ -261,6 +277,40 @@ class UnitTest(unittest.TestCase):
         self.args2.args_array2 = {
             "-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor"}
         self.proglock = ProgramLock(["cmdline"], "FlavorID")
+
+    @mock.patch("mongo_db_admin.gen_class.ArgParser")
+    def test_arg_parse2_true(self, mock_arg):
+
+        """Function:  test_arg_parse2_true
+
+        Description:  Test arg_parse2 returns false.
+
+        Arguments:
+
+        """
+
+        self.args.arg_parse2 = False
+
+        mock_arg.return_value = self.args
+
+        self.assertFalse(mongo_db_admin.main())
+
+    @mock.patch("mongo_db_admin.gen_libs.help_func")
+    @mock.patch("mongo_db_admin.gen_class.ArgParser")
+    def test_arg_parse2_true(self, mock_arg, mock_help):
+
+        """Function:  test_arg_parse2_true
+
+        Description:  Test arg_parse2 returns true.
+
+        Arguments:
+
+        """
+
+        mock_arg.return_value = self.args
+        mock_help.return_value = True
+
+        self.assertFalse(mongo_db_admin.main())
 
     @mock.patch("mongo_db_admin.gen_libs.help_func")
     @mock.patch("mongo_db_admin.gen_class.ArgParser")